+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Version label
 ### Changed
 - `connectProbe` method renamed to `attachToOrbit`
+- `triggerCrash` method renamed to `gameOver`
 ## [0.1] - 2026-07-24
 ### Added
 - Initial public release of SpinHook