+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - `connectProbe` method renamed to `attachToOrbit`
 - `triggerCrash` method renamed to `gameOver`
+- `README.md` updated
 ## [0.1] - 2026-07-24
 ### Added
 - Initial public release of SpinHook