+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Version label
+- Leaderboard
 ### Changed
 - `connectProbe` method renamed to `attachToOrbit`
 - `triggerCrash` method renamed to `gameOver`
+- `README.md` updated
 ## [0.1] - 2026-07-24
 ### Added
 - Initial public release of SpinHook